make server fns framework agnostic
