+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
   "leptos_macro",
   "leptos_reactive",
   "leptos_server",
+  "server_fn",
 
   # integrations
   "integrations/actix",
@@ -29,6 +30,8 @@ leptos_dom = { path = "./leptos_dom", default-features = false, version = "0.2.0
 leptos_macro = { path = "./leptos_macro", default-features = false, version = "0.2.0" }
 leptos_reactive = { path = "./leptos_reactive", default-features = false, version = "0.2.0" }
 leptos_server = { path = "./leptos_server", default-features = false, version = "0.2.0" }
+server_fn = { path = "./server_fn", default-features = false, version = "0.2.0" }
+server_fn_macro = { path = "./server_fn_macro", default-features = false, version = "0.2.0" }
 leptos_config = { path = "./leptos_config", default-features = false, version = "0.2.0" }
 leptos_router = { path = "./router", version = "0.2.0" }
 leptos_meta = { path = "./meta", default-feature = false, version = "0.2.0" }
