+++ leptos/Cargo.toml
@@ -17,6 +17,7 @@ leptos_server = { workspace = true }
 leptos_config = { workspace = true }
 tracing = "0.1"
 typed-builder = "0.12"
+server_fn = { workspace = true }
 
 [dev-dependencies]
 leptos = { path = ".", default-features = false }
