+++ leptos/src/lib.rs
@@ -168,6 +168,7 @@ pub use leptos_server::{
     self, create_action, create_multi_action, create_server_action,
     create_server_multi_action, Action, MultiAction, ServerFn, ServerFnError,
 };
+pub use server_fn;
 pub use typed_builder;
 mod error_boundary;
 pub use error_boundary::*;
