+++ leptos_macro/Cargo.toml
@@ -24,6 +24,7 @@ syn-rsx = "0.9"
 leptos_dom = { workspace = true }
 leptos_reactive = { workspace = true }
 leptos_server = { workspace = true }
+server_fn_macro = { workspace = true }
 convert_case = "0.6.0"
 uuid = { version = "1", features = ["v4"] }
