+++ leptos_server/Cargo.toml
@@ -11,6 +11,7 @@ readme = "../README.md"
 [dependencies]
 leptos_dom = { workspace = true }
 leptos_reactive = { workspace = true }
+server_fn = { workspace = true }
 form_urlencoded = "1"
 gloo-net = "0.2"
 js-sys = "0.3"
@@ -41,6 +42,7 @@ hydrate = [
 ssr = [
   #"leptos/ssr",
   "leptos_reactive/ssr",
+  "server_fn/ssr",
 ]
 stable = [
   #"leptos/stable",
