+++ leptos_server/src/action.rs
@@ -1,8 +1,9 @@
-use crate::{ServerFn, ServerFnError};
+use crate::{LeptosServerFnRegistry, ServerFnError};
 use leptos_reactive::{
     create_rw_signal, signal_prelude::*, spawn_local, store_value, ReadSignal,
     RwSignal, Scope, StoredValue,
 };
+use server_fn::ServerFn;
 use std::{future::Future, pin::Pin, rc::Rc};
 
 /// An action synchronizes an imperative `async` call to the synchronous reactive system.
@@ -111,7 +112,9 @@ where
 
     /// Associates the URL of the given server function with this action.
     /// This enables integration with the `ActionForm` component in `leptos_router`.
-    pub fn using_server_fn<T: ServerFn>(self) -> Self {
+    pub fn using_server_fn<T: ServerFn<Scope, LeptosServerFnRegistry>>(
+        self,
+    ) -> Self {
         let prefix = T::prefix();
         self.0.update_value(|state| {
             state.url = if prefix.is_empty() {
@@ -320,7 +323,7 @@ pub fn create_server_action<S>(
     cx: Scope,
 ) -> Action<S, Result<S::Output, ServerFnError>>
 where
-    S: Clone + ServerFn,
+    S: Clone + ServerFn<Scope, LeptosServerFnRegistry>,
 {
     #[cfg(feature = "ssr")]
     let c = move |args: &S| S::call_fn(args.clone(), cx);
