+++ leptos_server/src/lib.rs
@@ -79,14 +79,9 @@
 //!   or response or other server-only dependencies, but it does *not* have access to reactive state that exists in the client.
 
 use leptos_reactive::*;
-use proc_macro2::{Literal, TokenStream};
-use quote::TokenStreamExt;
-use serde::{de::DeserializeOwned, Deserialize, Serialize};
-use std::{future::Future, pin::Pin, str::FromStr};
-use syn::{
-    parse::{Parse, ParseStream},
-    parse_quote,
-};
+use serde::{Deserialize, Serialize};
+use server_fn::ServerFunctionRegistry;
+pub use server_fn::{Payload, ServerFn, ServerFnError};
 use thiserror::Error;
 
 mod action;
@@ -100,27 +95,77 @@ use std::{
 };
 
 #[cfg(any(feature = "ssr", doc))]
-type ServerFnTraitObj = dyn Fn(
-        Scope,
-        &[u8],
-    ) -> Pin<Box<dyn Future<Output = Result<Payload, ServerFnError>>>>
-    + Send
-    + Sync;
+type ServerFnTraitObj = server_fn::ServerFnTraitObj<Scope>;
 
 #[cfg(any(feature = "ssr", doc))]
 lazy_static::lazy_static! {
     static ref REGISTERED_SERVER_FUNCTIONS: Arc<RwLock<HashMap<&'static str, Arc<ServerFnTraitObj>>>> = Default::default();
 }
 
-/// A dual type to hold the possible Response datatypes
-#[derive(Debug)]
-pub enum Payload {
-    ///Encodes Data using CBOR
-    Binary(Vec<u8>),
-    ///Encodes data in the URL
-    Url(String),
-    ///Encodes Data using Json
-    Json(String),
+/// The registry of all Leptos server functions.
+pub struct LeptosServerFnRegistry;
+
+#[cfg(not(any(feature = "ssr", doc)))]
+impl ServerFunctionRegistry<Scope> for LeptosServerFnRegistry {}
+
+#[cfg(any(feature = "ssr", doc))]
+impl ServerFunctionRegistry<Scope> for LeptosServerFnRegistry {
+    type Error = ServerRegistrationFnError;
+
+    fn register(
+        url: &'static str,
+        server_function: Arc<ServerFnTraitObj>,
+    ) -> Result<(), Self::Error> {
+        // store it in the hashmap
+        let mut write = REGISTERED_SERVER_FUNCTIONS
+            .write()
+            .map_err(|e| ServerRegistrationFnError::Poisoned(e.to_string()))?;
+        let prev = write.insert(url, server_function);
+
+        // if there was already a server function with this key,
+        // return Err
+        match prev {
+            Some(_) => {
+                Err(ServerRegistrationFnError::AlreadyRegistered(format!(
+                    "There was already a server function registered at {:?}. \
+                     This can happen if you use the same server function name \
+                     in two different modules
+                on `stable` or in `release` mode.",
+                    url
+                )))
+            }
+            None => Ok(()),
+        }
+    }
+
+    /// Returns the server function registered at the given URL, or `None` if no function is registered at that URL.
+    fn get(url: &str) -> Option<Arc<ServerFnTraitObj>> {
+        REGISTERED_SERVER_FUNCTIONS
+            .read()
+            .ok()
+            .and_then(|fns| fns.get(url).cloned())
+    }
+
+    /// Returns a list of all registered server functions.
+    fn paths_registered() -> Vec<&'static str> {
+        REGISTERED_SERVER_FUNCTIONS
+            .read()
+            .ok()
+            .map(|fns| fns.keys().cloned().collect())
+            .unwrap_or_default()
+    }
+}
+
+#[cfg(any(feature = "ssr", doc))]
+/// Errors that can occur when registering a server function.
+#[derive(Error, Debug, Clone, Serialize, Deserialize)]
+pub enum ServerRegistrationFnError {
+    /// The server function is already registered.
+    #[error("The server function {0} is already registered")]
+    AlreadyRegistered(String),
+    /// The server function registry is poisoned.
+    #[error("The server function registry is poisoned: {0}")]
+    Poisoned(String),
 }
 
 /// Attempts to find a server function registered at the given path.
@@ -168,295 +213,11 @@ pub enum Payload {
 /// ```
 #[cfg(any(feature = "ssr", doc))]
 pub fn server_fn_by_path(path: &str) -> Option<Arc<ServerFnTraitObj>> {
-    REGISTERED_SERVER_FUNCTIONS
-        .read()
-        .ok()
-        .and_then(|fns| fns.get(path).cloned())
+    server_fn::server_fn_by_path::<Scope, LeptosServerFnRegistry>(path)
 }
 
 /// Returns the set of currently-registered server function paths, for debugging purposes.
 #[cfg(any(feature = "ssr", doc))]
 pub fn server_fns_by_path() -> Vec<&'static str> {
-    REGISTERED_SERVER_FUNCTIONS
-        .read()
-        .map(|vals| vals.keys().copied().collect())
-        .unwrap_or_default()
-}
-
-/// Holds the current options for encoding types.
-/// More could be added, but they need to be serde
-#[derive(Debug, PartialEq)]
-pub enum Encoding {
-    /// A Binary Encoding Scheme Called Cbor
-    Cbor,
-    /// The Default URL-encoded encoding method
-    Url,
-}
-
-impl FromStr for Encoding {
-    type Err = ();
-
-    fn from_str(input: &str) -> Result<Encoding, Self::Err> {
-        match input {
-            "URL" => Ok(Encoding::Url),
-            "Cbor" => Ok(Encoding::Cbor),
-            _ => Err(()),
-        }
-    }
-}
-
-impl quote::ToTokens for Encoding {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        let option: syn::Ident = match *self {
-            Encoding::Cbor => parse_quote!(Cbor),
-            Encoding::Url => parse_quote!(Url),
-        };
-        let expansion: syn::Ident = syn::parse_quote! {
-          Encoding::#option
-        };
-        tokens.append(expansion);
-    }
-}
-
-impl Parse for Encoding {
-    fn parse(input: ParseStream) -> syn::Result<Self> {
-        let variant_name: String = input.parse::<Literal>()?.to_string();
-
-        // Need doubled quotes because variant_name doubles it
-        match variant_name.as_ref() {
-            "\"Url\"" => Ok(Self::Url),
-            "\"Cbor\"" => Ok(Self::Cbor),
-            _ => panic!("Encoding Not Found"),
-        }
-    }
-}
-
-/// Defines a "server function." A server function can be called from the server or the client,
-/// but the body of its code will only be run on the server, i.e., if a crate feature `ssr` is enabled.
-///
-/// (This follows the same convention as the Leptos framework's distinction between `ssr` for server-side rendering,
-/// and `csr` and `hydrate` for client-side rendering and hydration, respectively.)
-///
-/// Server functions are created using the `server` macro.
-///
-/// The function should be registered by calling `ServerFn::register()`. The set of server functions
-/// can be queried on the server for routing purposes by calling [server_fn_by_path].
-///
-/// Technically, the trait is implemented on a type that describes the server function's arguments.
-pub trait ServerFn
-where
-    Self: Serialize + DeserializeOwned + Sized + 'static,
-{
-    /// The return type of the function.
-    type Output: Serialize;
-
-    /// URL prefix that should be prepended by the client to the generated URL.
-    fn prefix() -> &'static str;
-
-    /// The path at which the server function can be reached on the server.
-    fn url() -> &'static str;
-
-    /// The path at which the server function can be reached on the server.
-    fn encoding() -> Encoding;
-
-    /// Runs the function on the server.
-    #[cfg(any(feature = "ssr", doc))]
-    fn call_fn(
-        self,
-        cx: Scope,
-    ) -> Pin<Box<dyn Future<Output = Result<Self::Output, ServerFnError>>>>;
-
-    /// Runs the function on the client by sending an HTTP request to the server.
-    #[cfg(any(not(feature = "ssr"), doc))]
-    fn call_fn_client(
-        self,
-        cx: Scope,
-    ) -> Pin<Box<dyn Future<Output = Result<Self::Output, ServerFnError>>>>;
-
-    /// Registers the server function, allowing the server to query it by URL.
-    #[cfg(any(feature = "ssr", doc))]
-    fn register() -> Result<(), ServerFnError> {
-        // create the handler for this server function
-        // takes a String -> returns its async value
-
-        let run_server_fn = Arc::new(|cx: Scope, data: &[u8]| {
-            // decode the args
-            let value = match Self::encoding() {
-                Encoding::Url => serde_urlencoded::from_bytes(data)
-                    .map_err(|e| ServerFnError::Deserialization(e.to_string())),
-                Encoding::Cbor => ciborium::de::from_reader(data)
-                    .map_err(|e| ServerFnError::Deserialization(e.to_string())),
-            };
-            Box::pin(async move {
-                let value: Self = match value {
-                    Ok(v) => v,
-                    Err(e) => return Err(e),
-                };
-
-                // call the function
-                let result = match value.call_fn(cx).await {
-                    Ok(r) => r,
-                    Err(e) => return Err(e),
-                };
-
-                // serialize the output
-                let result = match Self::encoding() {
-                    Encoding::Url => match serde_json::to_string(&result)
-                        .map_err(|e| {
-                            ServerFnError::Serialization(e.to_string())
-                        }) {
-                        Ok(r) => Payload::Url(r),
-                        Err(e) => return Err(e),
-                    },
-                    Encoding::Cbor => {
-                        let mut buffer: Vec<u8> = Vec::new();
-                        match ciborium::ser::into_writer(&result, &mut buffer)
-                            .map_err(|e| {
-                                ServerFnError::Serialization(e.to_string())
-                            }) {
-                            Ok(_) => Payload::Binary(buffer),
-                            Err(e) => return Err(e),
-                        }
-                    }
-                };
-
-                Ok(result)
-            })
-                as Pin<Box<dyn Future<Output = Result<Payload, ServerFnError>>>>
-        });
-
-        // store it in the hashmap
-        let mut write = REGISTERED_SERVER_FUNCTIONS
-            .write()
-            .map_err(|e| ServerFnError::Registration(e.to_string()))?;
-        let prev = write.insert(Self::url(), run_server_fn);
-
-        // if there was already a server function with this key,
-        // return Err
-        match prev {
-            Some(_) => Err(ServerFnError::Registration(format!(
-                "There was already a server function registered at {:?}. This \
-                 can happen if you use the same server function name in two \
-                 different modules
-                on `stable` or in `release` mode.",
-                Self::url()
-            ))),
-            None => Ok(()),
-        }
-    }
-}
-
-/// Type for errors that can occur when using server functions.
-#[derive(Error, Debug, Clone, Serialize, Deserialize)]
-pub enum ServerFnError {
-    /// Error while trying to register the server function (only occurs in case of poisoned RwLock).
-    #[error("error while trying to register the server function: {0}")]
-    Registration(String),
-    /// Occurs on the client if there is a network error while trying to run function on server.
-    #[error("error reaching server to call server function: {0}")]
-    Request(String),
-    /// Occurs when there is an error while actually running the function on the server.
-    #[error("error running server function: {0}")]
-    ServerError(String),
-    /// Occurs on the client if there is an error deserializing the server's response.
-    #[error("error deserializing server function results {0}")]
-    Deserialization(String),
-    /// Occurs on the client if there is an error serializing the server function arguments.
-    #[error("error serializing server function results {0}")]
-    Serialization(String),
-    /// Occurs on the server if there is an error deserializing one of the arguments that's been sent.
-    #[error("error deserializing server function arguments {0}")]
-    Args(String),
-    /// Occurs on the server if there's a missing argument.
-    #[error("missing argument {0}")]
-    MissingArg(String),
-}
-
-/// Executes the HTTP call to call a server function from the client, given its URL and argument type.
-#[cfg(not(feature = "ssr"))]
-pub async fn call_server_fn<T>(
-    url: &str,
-    args: impl ServerFn,
-    enc: Encoding,
-) -> Result<T, ServerFnError>
-where
-    T: serde::Serialize + serde::de::DeserializeOwned + Sized,
-{
-    use ciborium::ser::into_writer;
-    use js_sys::Uint8Array;
-    use serde_json::Deserializer as JSONDeserializer;
-
-    #[derive(Debug)]
-    enum Payload {
-        Binary(Vec<u8>),
-        Url(String),
-    }
-    let args_encoded = match &enc {
-        Encoding::Url => Payload::Url(
-            serde_urlencoded::to_string(&args)
-                .map_err(|e| ServerFnError::Serialization(e.to_string()))?,
-        ),
-        Encoding::Cbor => {
-            let mut buffer: Vec<u8> = Vec::new();
-            into_writer(&args, &mut buffer)
-                .map_err(|e| ServerFnError::Serialization(e.to_string()))?;
-            Payload::Binary(buffer)
-        }
-    };
-
-    let content_type_header = match &enc {
-        Encoding::Url => "application/x-www-form-urlencoded",
-        Encoding::Cbor => "application/cbor",
-    };
-
-    let accept_header = match &enc {
-        Encoding::Url => "application/x-www-form-urlencoded",
-        Encoding::Cbor => "application/cbor",
-    };
-
-    let resp = match args_encoded {
-        Payload::Binary(b) => {
-            let slice_ref: &[u8] = &b;
-            let js_array = Uint8Array::from(slice_ref).buffer();
-            gloo_net::http::Request::post(url)
-                .header("Content-Type", content_type_header)
-                .header("Accept", accept_header)
-                .body(js_array)
-                .send()
-                .await
-                .map_err(|e| ServerFnError::Request(e.to_string()))?
-        }
-        Payload::Url(s) => gloo_net::http::Request::post(url)
-            .header("Content-Type", content_type_header)
-            .header("Accept", accept_header)
-            .body(s)
-            .send()
-            .await
-            .map_err(|e| ServerFnError::Request(e.to_string()))?,
-    };
-
-    // check for error status
-    let status = resp.status();
-    if (500..=599).contains(&status) {
-        return Err(ServerFnError::ServerError(resp.status_text()));
-    }
-
-    if enc == Encoding::Cbor {
-        let binary = resp
-            .binary()
-            .await
-            .map_err(|e| ServerFnError::Deserialization(e.to_string()))?;
-
-        ciborium::de::from_reader(binary.as_slice())
-            .map_err(|e| ServerFnError::Deserialization(e.to_string()))
-    } else {
-        let text = resp
-            .text()
-            .await
-            .map_err(|e| ServerFnError::Deserialization(e.to_string()))?;
-
-        let mut deserializer = JSONDeserializer::from_str(&text);
-        T::deserialize(&mut deserializer)
-            .map_err(|e| ServerFnError::Deserialization(e.to_string()))
-    }
+    server_fn::server_fns_by_path::<Scope, LeptosServerFnRegistry>()
 }
