+++ server_fn/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "server_fn"
+version = { workspace = true }
+edition = "2021"
+authors = ["Greg Johnston"]
+license = "MIT"
+repository = "https://github.com/leptos-rs/leptos"
+description = "RPC for any web framework."
+readme = "../README.md"
+
+[dependencies]
+form_urlencoded = "1"
+gloo-net = "0.2"
+js-sys = "0.3"
+lazy_static = "1"
+serde = { version = "1", features = ["derive"] }
+serde_urlencoded = "0.7"
+thiserror = "1"
+serde_json = "1.0.89"
+quote = "1"
+syn = { version = "1", features = ["full", "parsing", "extra-traits"] }
+proc-macro2 = "1"
+cfg-if = "1"
+ciborium = "0.2.0"
+xxhash-rust = { version = "0.8.6", features = ["const_xxh64"] }
+const_format = "0.2.30"
+
+[features]
+ssr = []
